move `polynomial_approximations` to `qualtran.linalg`
