+++ qualtran/linalg/polynomial_approximations.py
